# docs/swagger.yaml
definitions:
...
  dao.ServiceDetail:
    properties:
      access_control:
        $ref: '#/definitions/dao.AccessControl'
        type: object
      grpc_rule:
        $ref: '#/definitions/dao.GrpcRule'
        type: object
      http_rule:
        $ref: '#/definitions/dao.HttpRule'
        type: object
      info:
        $ref: '#/definitions/dao.ServiceInfo'
        type: object
      load_balance:
        $ref: '#/definitions/dao.LoadBalance'
        type: object
      tcp_rule:
        $ref: '#/definitions/dao.TcpRule'
        type: object
    type: object
...
  dto.ServiceAddGrpcInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      header_transfor:
        type: string
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceAddHTTPInput:
    properties:
      black_list:
        description: 黑名单ip
        type: string
      clientip_flow_limit:
        description: "\b客户端ip限流"
        type: integer
      header_transfor:
        description: header转换
        type: string
      ip_list:
        description: ip列表
        type: string
      need_https:
        description: 支持https
        type: integer
      need_strip_uri:
        description: 启用strip_uri
        type: integer
      need_websocket:
        description: 是否支持websocket
        type: integer
      open_auth:
        description: 关键词
        type: integer
      round_type:
        description: 轮询方式
        type: integer
      rule:
        description: 域名或者前缀
        type: string
      rule_type:
        description: 接入类型
        type: integer
      service_desc:
        description: 服务描述
        type: string
      service_flow_limit:
        description: 服务端限流
        type: integer
      service_name:
        description: 服务名
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        type: integer
      url_rewrite:
        description: url重写功能
        type: string
      weight_list:
        description: "\b权重列表"
        type: string
      white_list:
        description: 白名单ip
        type: string
    required:
    - ip_list
    - rule
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceAddTcpInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      header_transfor:
        type: string
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
...
  dto.ServiceListOutput:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/dto.ServiceListItemOutput'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  dto.ServiceStatOutput:
    properties:
      today:
        description: 列表
        items:
          type: integer
        type: array
      yesterday:
        description: 列表
        items:
          type: integer
        type: array
    type: object
...
  dto.ServiceUpdateTcpInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      id:
        type: integer
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - id
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
...
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /service/service_add_grpc:
    post:
      consumes:
      - application/json
      description: grpc服务添加
      operationId: /service/service_add_grpc
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddGrpcInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: grpc服务添加
      tags:
      - 服务管理
  /service/service_add_http:
    post:
      consumes:
      - application/json
      description: 添加HTTP服务
      operationId: /service/service_add_http
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddHTTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 添加HTTP服务
      tags:
      - 服务管理
  /service/service_add_tcp:
    post:
      consumes:
      - application/json
      description: tcp服务添加
      operationId: /service/service_add_tcp
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddTcpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: tcp服务添加
      tags:
      - 服务管理
  /service/service_delete:
    get:
      consumes:
      - application/json
      description: 服务删除
      operationId: /service/service_delete
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 服务删除
      tags:
      - 服务管理
  /service/service_detail:
    get:
      consumes:
      - application/json
      description: 服务详情
      operationId: /service/service_detail
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dao.ServiceDetail'
              type: object
      summary: 服务详情
      tags:
      - 服务管理
  /service/service_list:
    get:
      consumes:
      - application/json
      description: 服务列表
      operationId: /service/service_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page_no
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceListOutput'
              type: object
      summary: 服务列表
      tags:
      - 服务管理
  /service/service_stat:
    get:
      consumes:
      - application/json
      description: 服务统计
      operationId: /service/service_stat
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ServiceStatOutput'
              type: object
      summary: 服务统计
      tags:
      - 服务管理
  /service/service_update_grpc:
    post:
      consumes:
      - application/json
      description: grpc服务更新
      operationId: /service/service_update_grpc
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateGrpcInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: grpc服务更新
      tags:
      - 服务管理
  /service/service_update_http:
    post:
      consumes:
      - application/json
      description: 修改HTTP服务
      operationId: /service/service_update_http
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateHTTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改HTTP服务
      tags:
      - 服务管理
  /service/service_update_tcp:
    post:
      consumes:
      - application/json
      description: tcp服务更新
      operationId: /service/service_update_tcp
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateTcpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: tcp服务更新
      tags:
      - 服务管理
swagger: "2.0"